+++ custom_components/yahoo_jp_weather/const.py
@@ -3,6 +3,7 @@
 from datetime import timedelta
 
 DOMAIN = "yahoo_jp_weather"
+CONF_ENTITY_UNIQUE_ID = "entity_unique_id"
 CONF_NAME = "name"
 DEFAULT_NAME = "江戸川区"
 DEFAULT_URL = "https://weather.yahoo.co.jp/weather/jp/13/4410/13123.html"